reusable <HelpArticleLayout> component for React to wrap all articles consistently
