+++ src/App.jsx
@@ -13,16 +13,16 @@ import CompleteProfile from "./Pages/CompleteProfile";
 import CookiesPolicy from "./Pages/CookiesPolicy";
 import EditRide from "./Pages/EditRide";
 import Help from "./Pages/Help/Help";
+import AccountSecurityHelp from "./Pages/Help/YourProfileAndAccount/AccountSecurity";
 import CreatingAccount from "./Pages/Help/YourProfileAndAccount/CreatingAccount";
 import AccountHelp from "./Pages/Help/YourProfileAndAccount/HelpAccount";
 import RatingsHelp from "./Pages/Help/YourProfileAndAccount/HelpRatings";
 import LoginTroubleshooting from "./Pages/Help/YourProfileAndAccount/HelpTroubleshooting";
 import ManageProfile from "./Pages/Help/YourProfileAndAccount/ManageProfile";
-import AccountSecurityHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/AccountSecurity";
-import DeleteProfileHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/DeteleProfile";
-import EditProfileHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/EditProfile";
+import DeleteProfileHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/FilesInside/DeteleProfile";
+import EditProfileHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/FilesInside/EditProfile";
+import NotificationsHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/FilesInside/Notifications";
 import ManageAccountOverview from "./Pages/Help/YourProfileAndAccount/Managing Your Account/ManagingYourAccount";
-import NotificationsHelp from "./Pages/Help/YourProfileAndAccount/Managing Your Account/Notifications";
 import Homepage from "./Pages/Homepage";
 import IndividualRide from "./Pages/IndividualRide";
 import Login from "./Pages/Login";
