--- src/Pages/Help/StylesHelp/HelpYourAccount.css
@@ -1,49 +1,0 @@
-.help-container {
-  max-width: 800px;
-  margin: 3rem auto;
-  padding: 2rem;
-  background: #f9f9fb;
-  border-radius: 16px;
-  box-shadow: 0 4px 14px rgba(0, 0, 0, 0.05);
-  font-family: "Segoe UI", sans-serif;
-}
-
-.help-heading {
-  font-size: 2rem;
-  font-weight: 700;
-  margin-bottom: 1.5rem;
-  color: #2c3e50;
-}
-
-.help-section {
-  margin-top: 2rem;
-}
-
-.help-subheading {
-  font-size: 1.4rem;
-  font-weight: 600;
-  margin-bottom: 1rem;
-  color: #34495e;
-}
-
-.help-list {
-  list-style-type: disc;
-  padding-left: 1.5rem;
-  line-height: 1.7;
-  color: #555;
-}
-
-.help-list li {
-  margin-bottom: 1.2rem;
-}
-
-.help-contact {
-  margin-top: 2rem;
-  font-weight: 500;
-  color: #2c3e50;
-}
-
-.help-contact a {
-  color: #0077cc;
-  text-decoration: underline;
-}
