+++ src/Pages/Help/YourProfileAndAccount/CreatingAccount.jsx
@@ -1,39 +1,30 @@
 import React from "react";
-import { Link } from "react-router-dom";
-import Breadcrumb from "../Breadcrumb";
+import HelpArticleLayout from "../HelpArticleLayout";
 
 export default function CreatingAccount() {
-  const breadcrumbItems = [
+  const breadcrumb = [
     { label: "Help", to: "/help" },
     { label: "Account", to: "/help/account" },
     { label: "Creating an Account" },
   ];
 
   return (
-    <div className="help-wrapper">
-      <Breadcrumb items={breadcrumbItems} />
-
-      <header className="help-header">
-        <h1>Creating an Account</h1>
-        <p className="subtext">
-          Learn how to get started with your ShareRide profile.
-        </p>
-      </header>
-
-      <section className="help-section">
-        <p>To create an account:</p>
-        <ol>
-          <li>
-            Go to the <Link to="/register">registration page</Link>.
-          </li>
-          <li>Fill in your name, email, and create a password.</li>
-          <li>Click "Sign Up" and check your email for a verification link.</li>
-          <li>Once verified, you can log in and complete your profile.</li>
-        </ol>
-        <p>
-          Need help? <Link to="/help/contact">Contact support</Link>.
-        </p>
-      </section>
-    </div>
+    <HelpArticleLayout
+      title="Creating an Account"
+      description="Learn how to get started with your ShareRide profile."
+      breadcrumb={breadcrumb}
+    >
+      <ol>
+        <li>
+          Go to the <a href="/register">registration page</a>.
+        </li>
+        <li>Fill in your name, email, and create a password.</li>
+        <li>Click "Sign Up" and check your email for a verification link.</li>
+        <li>Once verified, you can log in and complete your profile.</li>
+      </ol>
+      <p>
+        Need help? <a href="/contact-support">Contact support</a>.
+      </p>
+    </HelpArticleLayout>
   );
 }
