--- src/Pages/Help/YourProfileAndAccount/Managing Your Account/DeteleProfile.jsx
@@ -1,57 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import Breadcrumb from "../../Breadcrumb";
-import "../../StylesHelp/Help.css";
-
-export default function DeleteProfileHelp() {
-  const breadcrumbItems = [
-    { label: "Help", to: "/help" },
-    { label: "Account", to: "/help/account" },
-    {
-      label: "Managing Your Account",
-      to: "/help/account/managing-your-account",
-    },
-    { label: "Deleting Your Profile" },
-  ];
-
-  return (
-    <div className="help-wrapper">
-      <Breadcrumb items={breadcrumbItems} />
-
-      <header className="help-header">
-        <h1>Deleting Your Profile</h1>
-        <p className="subtext">
-          Need to leave ShareRide? Here's how to delete your account.
-        </p>
-      </header>
-
-      <section className="help-section">
-        <h2>Important before deleting</h2>
-        <ul>
-          <li>
-            All your ride history and messages will be permanently removed.
-          </li>
-          <li>You will lose access to your account and booked rides.</li>
-        </ul>
-
-        <h2>How to delete your profile</h2>
-        <ol>
-          <li>Log in to your account.</li>
-          <li>
-            Go to <strong>Account Settings</strong> &gt;{" "}
-            <strong>Privacy</strong>.
-          </li>
-          <li>
-            Click <strong>Delete My Account</strong>.
-          </li>
-          <li>Confirm the action when prompted.</li>
-        </ol>
-
-        <p>
-          Changed your mind? You can always contact{" "}
-          <Link to="/help/contact">Support</Link> before deletion.
-        </p>
-      </section>
-    </div>
-  );
-}
