--- src/Pages/Help/YourProfileAndAccount/Managing Your Account/Notifications.jsx
@@ -1,52 +1,0 @@
-import React from "react";
-import Breadcrumb from "../../Breadcrumb"; // adjust path if needed
-import "../../StylesHelp/Help.css";
-
-export default function NotificationsHelp() {
-  const breadcrumbItems = [
-    { label: "Help", to: "/help" },
-    { label: "Account", to: "/help/account" },
-    {
-      label: "Managing Your Account",
-      to: "/help/account/managing-your-account",
-    },
-    { label: "Notifications & Communication" },
-  ];
-
-  return (
-    <div className="help-wrapper">
-      <Breadcrumb items={breadcrumbItems} />
-
-      <header className="help-header">
-        <h1>Notifications & Communication</h1>
-        <p className="subtext">
-          Manage how you receive updates about rides and account activity.
-        </p>
-      </header>
-
-      <section className="help-section">
-        <h2>Where to manage notifications</h2>
-        <ol>
-          <li>
-            Log in to your account and go to <strong>Account Settings</strong>.
-          </li>
-          <li>
-            Select the <strong>Notifications</strong> tab.
-          </li>
-          <li>
-            Toggle on/off email, SMS, or in-app notifications based on your
-            preference.
-          </li>
-        </ol>
-
-        <h2>Types of notifications</h2>
-        <ul>
-          <li>Ride reminders and status updates</li>
-          <li>Messages and chat activity</li>
-          <li>Account security and login alerts</li>
-          <li>News, promotions, and feature updates (optional)</li>
-        </ul>
-      </section>
-    </div>
-  );
-}
